requirements.txt
flask
pandas
python-dotenv
mgrs
packaging